mr fixes
