+++ Projects/Contracts/Sources/Screens/ContractInformation.swift
@@ -69,7 +69,7 @@ extension ContractInformation: Presentable {
 
         form.appendSpacing(.custom(20))
 
-        if contract.currentAgreement?.status != .terminated {
+        if contract.currentAgreement.status != .terminated {
             if Localization.Locale.currentLocale.market == .se {
                 let changeAddressButton = ButtonRowViewWrapper(
                     title: L10n.HomeTab.editingSectionChangeAddressLabel,
