+++ Projects/Contracts/Sources/Viewables/ContractTable.swift
@@ -62,7 +62,7 @@ extension ContractTable: Viewable {
                     .flatMap { $0.contracts }
             case .terminated:
                 return state.contracts.filter { contract in
-                    contract.currentAgreement?.status == .terminated
+                    contract.currentAgreement.status == .terminated
                 }
             case .none: return []
             }
